requirements.txt
absl-py==2.1.0
attrs==24.2.0
cffi==1.17.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
imageio==2.36.1
jax==0.4.37
jaxlib==0.4.36
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mediapipe==0.10.18
ml_dtypes==0.5.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==11.0.0
protobuf==4.25.5
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
scipy==1.14.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
